fix: route-specific middleware and improve logging
- Fixed critical bug where all middleware was applied globally
  - Middleware registered for specific routes now only applies to matching paths
  - Global middleware (with "*") correctly applies to all routes
  - Proper execution order: global first, then route-specific
  - Multiple middleware for same route are appended, not replaced
- Changed options.json not found message from WARN to INFO level
- Fixed all example compilation errors
- Updated documentation to reflect correct middleware behavior

BREAKING CHANGE: Route-specific middleware now works as intended.
Previously, all middleware was applied to all routes regardless of
the specified path pattern.

+++ examples/configuration/main.go
@@ -77,7 +77,7 @@ func runProgrammaticConfig() {
 
 	fmt.Println("Configuration:")
 	fmt.Printf("  Port: %s\n", server.Options.Addr)
-	fmt.Printf("  Rate Limit: %d req/s\n", server.Options.RateLimit)
+	fmt.Printf("  Rate Limit: %.0f req/s\n", float64(server.Options.RateLimit))
 	fmt.Printf("  Burst Limit: %d\n", server.Options.Burst)
 
 	fmt.Println("\nPress Enter to continue to next method...")
@@ -119,7 +119,7 @@ func runJSONConfig() {
 
 	fmt.Println("Configuration loaded from JSON:")
 	fmt.Printf("  Port: %s\n", server.Options.Addr)
-	fmt.Printf("  Rate Limit: %d req/s\n", server.Options.RateLimit)
+	fmt.Printf("  Rate Limit: %.0f req/s\n", float64(server.Options.RateLimit))
 	fmt.Printf("  Burst Limit: %d\n", server.Options.Burst)
 
 	// Cleanup
@@ -151,7 +151,7 @@ func runEnvConfig() {
 
 	fmt.Println("Configuration loaded from environment:")
 	fmt.Printf("  Port: %s\n", server.Options.Addr)
-	fmt.Printf("  Rate Limit: %d req/s\n", server.Options.RateLimit)
+	fmt.Printf("  Rate Limit: %.0f req/s\n", float64(server.Options.RateLimit))
 	fmt.Printf("  Burst Limit: %d\n", server.Options.Burst)
 
 	// Cleanup
